interpreter: Split Disabler from interpreterbase.py
